feat: add enpoint factory
