+++ src/config/http/middlewares/index.ts
@@ -1,0 +1,2 @@
+export * from "./AuthMiddleware";
+export * from "./AttachmentsParser";
