+++ src/routes/user/updateUser.ts
@@ -1,9 +1,9 @@
 import { services } from "app";
-import { taggedEndpointsFactory } from "config/http/EndpointFactory";
+import { endpointFactory } from "config/http/Endpoint";
 import { UpdateUserSchema, UserSchema } from "entities/user.entity";
 import { withMeta } from "express-zod-api";
 
-export const updateUser = taggedEndpointsFactory.build({
+export const updateUser = endpointFactory.tagged.build({
   tag: "users",
   method: "put",
   shortDescription: "Actualizacion de usuarios",
